% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine_version}
\alias{engine_version}
\title{Version of the engine}
\usage{
engine_version()
}
\value{
A list with `numerics`, the version of the numerical definition
  (features from different numerics versions are not comparable), and
  `crate`, the version of the Rust code.
}
\description{
Version of the engine
}
